--- app/graphql/mutations/monitor_resources_mutation.rb
@@ -1,13 +1,0 @@
-# frozen_string_literal: true
-
-module Mutations
-  class MonitorResourcesMutation < Mutations::BaseMutation
-    argument :input, Types::Custom::Inputs::Mutations::MonitorResourcesInput, required: true
-    type Boolean
-
-    def resolve(params)
-      super(params)
-      ::Services::Monitoring::HandleMonitorResourcesService.call(params: @params)
-    end
-  end
-end
