+++ app/graphql/types/mutation_type.rb
@@ -15,11 +15,11 @@ module Types
           description: 'Logout user'
 
     field :subscribe_user_to_newsletter,
-          mutation: Mutations::SubscribeUserToNewsletterMutation,
+          mutation: Mutations::Newsletter::SubscribeUserToNewsletterMutation,
           description: 'Save user to newsletter'
 
     field :unsubscribe_user_from_newsletter,
-          mutation: Mutations::UnsubscribeUserFromNewsletterMutation,
+          mutation: Mutations::Newsletter::UnsubscribeUserFromNewsletterMutation,
           description: 'Remove user from newsletter'
 
     field :update_user,
@@ -31,15 +31,15 @@ module Types
           description: "Remove user's account"
 
     field :add_opinion,
-          mutation: Mutations::AddOpinionMutation,
+          mutation: Mutations::Opinions::AddOpinionMutation,
           description: "Add user's opinion"
 
     field :add_order,
-          mutation: Mutations::AddOrderMutation,
+          mutation: Mutations::Orders::AddOrderMutation,
           description: "Add user's order"
 
     field :monitor_resources,
-          mutation: ::Mutations::MonitorResourcesMutation,
+          mutation: ::Mutations::Tools::MonitorResourcesMutation,
           description: "Publish resources's state on Kafka for monitoring purpose"
   end
 end
